backend/requirements.txt
notebook
numpy
pandas
nltk
scikit-learn
flask
joblib
opencv-python
pytesseract
pillow
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
Werkzeug==2.3.7
joblib==1.3.0
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0